Fix validation on copy card page (#279)
* Fix validation on copy card page

https://eaflood.atlassian.net/browse/WC-498

Our validation for the `temp_cards` field on the copy card form assumed that leaving it blank would sub in a zero, because the field is classed as an Integer in the database.

This is not the case - if the field already existed, the value is replaced with null, and if it doesn't exist then it's not created.

This means that when the order_items are created, checks against the value of the field fail and we get an error.

This commit modifies our validation to no longer allow blank values.

* If temp_cards is left blank, sub in 0

If the user submits an empty value, sub in 0 so validation passes and they can continue the renewal process.

* Deal with blank temp_card values when building Order

Order.new_order assumed that temp_cards will have a value. If it doesn't, then this causes the error which was initially reported.

Once all the renewals with blank temp_cards have passed through the system, this may not be necessary - so included a comment to remind us to review this in the future.