Rescue StandardErrors in request services (#284)
* Rescue StandardErrors in request services

Previously our services which make requests were able to rescue specific errors. However, the way each `rescue` block dealt with the errors was generally the same, causing duplication. This also meant that any errors which weren't specified were not rescued and instead caused disruption to the user.

This PR replaces duplicated, error-specific `rescue` blocks with a `rescue StandardError => e` block, which reduces duplication and covers a much broader range of errors.

* Make specs meore generic and remove duplication

Now that we don't have different `rescue` blocks for different errors, we don't need to have individual tests for all the different errors we handle.