Generate & attach PDF cert. to confirmation email (#273)
Generate & attach PDF cert. to confirmation email

https://eaflood.atlassian.net/browse/WC-479

We attach PDF certificates to new registration emails. However, for renewals, we currently ask them to log in.

If the user can't log in and renewed through an AD process instead, they have no way to access the certificate.

This change will generate and attach the updated certificate to the "renewal complete" email so no user needs to log in to access it.

It is a large change, so key things to point out are

Introduction of presenters
--------------------------
As we are just trying to replicate the existing PDF, we are looking to migrate across the existing certificate template. However in review we found it relied on complex logic in the view, and lots of 'helper' methods which are actually on the model rather than as view helpers.

With all this logic needed to generate the view, we decided the cleanest solution was to introduce presenters.

> Presenters give you an object oriented way to approach view helpers.

We leaned heavily on http://nithinbekal.com/posts/rails-presenters/ and http://railscasts.com/episodes/287-presenters-from-scratch for the basis of the implementation. There is a gem called Draper (https://github.com/drapergem/draper) that allows you to decorate models with presentation logic, but we didn't feel this required adding a dependency, plus we preferred the idea of presenters as something completely separate from the model.

Sped up email attachment tests
------------------------------

Initially we followed the pattern of placing our email attachment tests in individual blocks, each with their own setup. However this meant just this block of 5 tests was taking 10 seconds to run.

To resolve it we could have placed all the expectations into a single test, but then we would lose the clarity of seeing the expected behaviour. So instead we opted for a `before(:all)` block for this specific context. This runs just once, but each test can then access the variables it creates. In this context using the same certificate and email is just as valid as creating a fresh one for each test.

Ensure error in pdf does not block renewal
------------------------------------------

Though it won't be ideal for the user that they don't get their PDF, we don't want an error when trying to generate it and attach it to the email to block the renewal from completing. Hence we wrap the generation of the pdf in a rescue block, and return nil (and handle it) should any issues occur.

Remove pdf certificate layout
-----------------------------

Unlike the frontend project, we have built the template for the certificate in a single file with no layout or partials.

We felt having a layout felt like we were pre-empting reuse. Removing it removes this preemptiveness, simplifies things, and brings it inline with the existing mailer (which doesn't use a layout or partials).

Add our own version of wicked_pdf_image_tag
-------------------------------------------

Using Wicked PDF within an engine appears to mean that its helpers are not available to us when the engine is then mounted in a host app.

The one other reference to trying to do this suggests writing your own custom helper in place of using WickedPdf's (https://github.com/mileszs/wicked_pdf/issues/400).

Hence we add our own custom version of the helper, that points to a file within the engine.