Allow renewals that expire after submission to complete (#287)
* Allow renewals that expire after submission to complete

https://eaflood.atlassian.net/browse/WC-514

If a user submits a renewal request before the expiry date, but the renewal isn't confirmed until after the expiry date because of delayed payment or conviction checks, it should still be possible to complete the renewal.

* Add more tests for renewing an expired registration

These tests should pass once we have implented the desired behaviour.

* Allow renewal from an expired state

No expired registrations will meet the criteria set by our guard methods though.

* Refactor to use a single guard method for renew action

All guard clauses must return true for an AASM transition to be allowed, and our conditions are about to get more complicated. So enclosing our current checks in a single method won't affect current behaviour at all, but will make it easier for us to add more conditions.

* Allow renewal if the application was submitted but not immediate

Check if there is a matching TransientRegistration in the state of "renewal received". If there is, we know that their application to renew would not have been instant, as they only hit this page if there is a delay for payment or conviction checks. However, they must have met the other conditions for renewal when they began the process, as no TransientRegistration would have existed yet.

Note that we don't check for "renewal complete" as anything in this state would have been renewed immediately with no delay for NCCC checks.

* If application hasn't been submitted, no renewal for expired reg

A lot of our tests were failing because we'd assumed an expired registration would never be renewed. This is still true, with the one exception of registrations where the application has been submitted. So let's add an extra check to teh logic to confirm this.