Update rubocop config (#297)
The existing configuration for rubocop was taken from our [DST-Guides](https://github.com/DEFRA/dst-guides) repo, though unbeknownst to the dev that took them that version was already out of date.

Now its gotten to the stage that some linters are erroring when attempting to run rubocop checks whilst we work, which means dev's need to decide to either ignore the notifications or disable the linter; neither is a great decision.

So this updates the config to match the front and back office projects. It should be noted this does not resolve any violations. That is something we intend to resolve in the future.

With the update we can switch back on our linters and receive no more pesky error messages 😊