Fix conviction check to handle grace window (#296)
https://eaflood.atlassian.net/browse/WC-512

During testing of the grace window it was spotted that if a renewal was expired, and it matched during conviction checks, then it could not be continued because the button to access convictions approvals did not display.

This was because `TransientRegistration.pending_manual_conviction_check?` had a guard clause that checked the status was `ACTIVE`. This will no longer be the case for registrations which are EXPIRED but in the grace window, hence we change it to use `may_renew?`.

We also learnt doing this that `may_renew?` will only work if called on a `metaData` object that is linked to a registration, because of the call to `registration.identifier` in `CanChangeRegistrationStatus.renewal_application_submitted?`. Hence we had to tweak the tests and code to account for this.

Also included

- some fixes which resolve tests that failed locally where the grace window was set, that didn't fail in Travis-CI where it wasn't
- resolution of some rubocop violations (now rubocop is working again)