Allow expired renewal to complete post declaration (#300)
https://eaflood.atlassian.net/browse/WC-526

We know now that we have renewals linked to registrations that have expired during the payment phase.

In a lot of these cases its because Worldpay has failed, and we have asked NCCC to use the workaround of sending the renewal back to the payment summary page and then completing it as a BAC's payment. The problem we are now encountering in greater numbers is renewals are expiring before the missed payment can be spotted and the workaround enacted.

We have therefore agreed that once a renewal has had the declaration accepted, it can still be renewed even if it then expires. This change covers the amendments necessary to support this.

As part of doing this we also refactored `CanChangeRegistrationStatus` by bringing in a new `ExpiryCheckService` (based on the `ExpiryDateService` in the frontend). We originally thought the service contained functionality we'd need, but in retrospect that was not the case. However we felt the refactoring benefitted the project, so kept them in.

Finally there are minor changes to resolve rubocop infractions.