Improve dashboard search performance (#302)
https://eaflood.atlassian.net/browse/WC-531

In production our users have reported that when attempting to apply a filter in the dashboard, it becomes unresponsive and eventually leads to a "service unavailable" page.

Initial investigation has shown that the problem is each time a user searches the app it's pulling back all transient registrations and then iterating through them, checking for the matching criteria.

So this change is about improving the performance of the search in the dashboard, such that it no longer times out, and is a scalable solution i.e. we won't be back here again in another 6 months.

Mongoid has the functionality to add [Named scopes](https://docs.mongodb.com/mongoid/master/tutorials/mongoid-queries/#named-scopes).

> Named scopes are simply criteria defined at class load that are referenced by a provided name. Just like normal criteria, they are lazy and chainable.

So we've solved the puzzle by adding named scopes that mimic the current criteria in the back office `TransientRegistrationFinderService`. These will then be integrated with the search in the back office, and it's use of [Kaminari](https://github.com/kaminari/kaminari) for paging to limit the result size.