Add workflow states to conviction_sign_offs (#303)
* Add workflow states to conviction_sign_offs

https://eaflood.atlassian.net/browse/WC-483

The current implementation of the conviction checking workflow allows a flagged renewal to be checked, then approved or rejected. However, the actual process is more complex, and a renewal which does have matching convictions goes through an internal process before a final decision is made. This commit adds different states to the conviction_sign_off in order to better represent the workflow.

* Implement AASM for conviction_sign_offs

Create a new concern to manage the conviction workflow stages and include it in the ConvictionSignOff model. A conviction_sign_off always starts as :possible_match. At this point it can be approved, rejected or move into :checks_in_progress when more work is required. Something in :checks_in_progress can then be approved or rejected.

We temporarily refer to the approve event as 'sign off' instead so as not to clash with the approve method currently on the ConvictionSignOff model.

These names and stages may change after further consultation with NCCC.

* Add frozen_string_literal: true comments

If I touch it, I fix it.

* Update confirmed and confirmed_at as part of sign_off

Move the updates to these attributes out of the old approve method and into the new sign_off action.

This is part of getting rid of the old approve method entirely and relying on the state machine to handle all the work - moving confirmed_by will be slightly more complex, so we save that for the next commit.

* Update confirmed_by as part of sign_off

Pass the current_user to the sign_off action so we can update it there instead of in the old approve method.

We should now be able to remove the old approve method and replace it with sign_off.

* Remove old approve method

The new state machine sign_off action does everything we need, so this method is no longer needed.

* Rename sign_off action to approve

This is what we wanted to name it all along, but couldn't while the old approve method was still in place.

The back office will call `conviction_sign_off.approve` anyway but may need some minor tweaks, as it previously assumed that would persist the changes, which will now only happen with `.approve!`

* Revoke transient_registration when rejecting conviction_sign_off

Move this functionality out of the back office and into the state machine, similar to what we did for the approve action.

We use `_parent` instead of `transient_registration` to make this more flexible in case this process has to deal with registrations as well.