Add scopes for different conviction check states (#304)
* Add scopes for different conviction check states

https://eaflood.atlassian.net/browse/WC-483

Now that we have workflow_states for conviction_sign_offs, we can use these to filter transient_registrations. This PR creates new scopes for the various states a conviction_sign_off can be in.

* Add scope for conviction_sign_offs in :possible_match state

This scope should only return transient_registrations which have a conviction_sign_off with a workflow_state of :possible_match.

* Add scope for conviction_sign_offs in :checks_in_progress state

This scope should only return transient_registrations which have a conviction_sign_off with a workflow_state of :checks_in_progress.

* Add scope for conviction_sign_offs in :approved state

This scope should only return transient_registrations which have a conviction_sign_off with a workflow_state of :approved.

* Add scope for conviction_sign_offs in :rejected state

This scope should only return transient_registrations which have a conviction_sign_off with a workflow_state of :rejected.