Refactor scope tests for consistency (#305)
* Refactor scope tests for consistency

Update our specs for transient_registration scopes so we are testing them in a consistent way.

* Change 'context' to 'describe'

As these blocks describe the method, rather than the context in which it is being called.

* Use let to reduce duplication in test objects

Several scope specs were creating near-identical objects in the individual specs. We can reduce duplication by specifying these using `let`.

* Reuse existing test object for reg_identifier search spec

We can use any transient_registration with a unique reg_identifier to test this, so `in_progress_renewal` fits the bill.

* Move example objects for name search test into let blocks

This is a precursor to further refactoring, so we can test the expectations individually.

Also add some extra expectations that the results include the test objects, rather than just checking the length.

* Move postcode test objects into let block

Then test that the results include the matching object and do not include the non-matching one.

Also a precursor to testing expectations separately.

* Define scope for filter specs in describe block

Use `let` to define the scope we are testing once, just inside the `describe` block. This means the individual specs have less duplication and are easier to read.

* Define search_term scope once and change term in context

As we did with the filters, define the scope for `search_term` using a `let` block right inside the `define` block. We provide a default for term, but can use additional `let` blocks to modify this depending on the context.

* Split expectations into separate blocks

This means we can run tests at a more granular level, and means the format is more consistent with our other scope tests.

* Always define term just inside context block

A minor reordering, but makes the tests a bit easier to read.