Add missing relationship between metaData & trans reg (#316)
Taking the finance details for example, we define the relationship both in the `CanHaveRegistrationAttributes` concern and in `FinanceDetails` between `FinanceDetails` and `TransientRegistration`.

However we don't have this declaration on `MetaData` class for `TransientRegistration` even though we need it, so this change adds the missing declaration.

Obviously we have been fine up to now without it. However this is inconsistent with how we have done it elsewhere so we feel its important to make this change.