Refactor logic around can renew checks (#307)
https://eaflood.atlassian.net/browse/WC-533

WC-533 highlighted that we were doing checks for whether something can be renewed in 2 different places, in 2 different ways. We agreed to move the checks to one place, but additional review has highlighted that we're using our current checks for mixed purposes, and as such are not being very efficient.

Prior to this change `app/models/concerns/waste_carriers_engine/can_change_registration_status.rb` had `renewal_allowed?` which was hooked up to the AASM event `renew`. Hence within the code you could call `metaData.may_renew?`.

You would call the renew event when you wish to transition a registration to be 'ACTIVE`, and to extend the existing expiry date and update the activation timestamps. In essence you called this when you were completing a renewal.

However these changes are only part of what happens to a registration when you complete a renewal. Hence we have the `app/services/waste_carriers_engine/renewal_completion_service.rb`. This highlights there are inefficiencies. For example

- When you are ready to complete a renewal you instantiate a `RenewalCompletionService` passing in the transient registration. In the initializer it also hits the db to pull back the matching registration
- You then call `complete_renewal` which ends up calling `metaData.may_renew?`
- This doesn't have a reference to the transient registration, so makes another hit on the database for that
- It then carries out its checks and returns true or false
- If true we're back in `complete_renewal` making a series of changes including a call to `update_registration`
- In `update_registration` we end up again calling `metaData.renew` (in the guise of requesting a transition not just querying if we can) which in turn repeats the same checks we've already performed, and hits the db for a 3rd time to get the transient registration
- Also in the current form, `RenewalCompletionService` isn't checking if there are any pending payment or conviction checks, as these were being done by the back office.

Hence the changes here do 2 things

- split out the checks into the correct contexts; can I start a renewal and continue a renewal, and can I complete a renewal
- make the checks more efficient such that we reduce the number of calls to the db for data we already have

Because of the size of the refactoring and the number of files touched there are also some additional changes to resolve rubocop violations, and changes to the tests and factories.