Add stuck? method to the transient registration (#327)
* Add stuck? method to the transient registration

https://eaflood.atlassian.net/browse/WC-533

As part of helping NCCC manage renewals that get stuck we are adding a new status to the dashboard in [Waste carriers back office PR 217](https://github.com/DEFRA/waste-carriers-back-office/pull/217).

To display the status we need to know if a transient registration is stuck. Initial designs rely on too much logic in the view, or falling back on helpers that would try and make the determination based on existing properties and methods.

So instead we are adding a method to the transient registration so we can simply ask it directly; are you stuck?