Refactor specs around renewal checks (#326)
* Refactor specs around renewal checks

One of the transient_registration specs was passing on Travis and some machines, but failing locally for me. After some investigation I discovered this was because the attempts to stub `Rails.configuration.grace_window` were failing, because that value was being called by the ExpiryCheckService, not the model being tested. Without the stub, the environment variable was always used, resulting in a failing test.

The solution to this issue is to make sure that the model spec is responsible for testing the model, and the service spec is responsible for testing the service. This PR refactors some specs and moves some others to achieve this.

* Stub ExpiryCheckService methods checking windows in model spec

This refactor switches from modifying values used by the ExpiryCheckService to stubbing its responses.

This means the model spec no longer has to know about the inner workings of the ExpiryCheckService (as the model itself doesn't know these things) and only has to deal with what it returns.

* Reverse order of tests to match order in the method

This makes the spec for the model follow the checks in the method a little more closely.

* Move BST/GMT checks from model spec to service spec

The service spec already covers how `expiry_date` is initially set, but this means we have some more coverage of the knock-on effects, so I think the specs are worth moving rather than deleting.

* Add spec for #expiry_date_after_renewal

Running this spec file on its own and looking at the test coverage revealed we were never actually testing it directly.

* Minor spacing tweaks