Fix style issues in controller and request specs (#335)
* Fix style issues in controller and request specs

This PR fixes up style issues in our controller and request specs.

The actual controllers have been dealt with in https://github.com/DEFRA/waste-carriers-renewals/pull/334

* Add frozen_string_literal comments

* Don't name form param when passing it to shared examples

We use shared examples heavily for GET and POST requests in our form pages. Previously we named each parameter we passed to the example, mostly for ease of reading - however, this triggers the 'useless assignment to variable' cop.

The form name is fairly self-explanatory, so we remove the assignment for that. However, other options will require further work.

* Pass options hash to shared example for POST forms

Rather than passing in a bunch of named arguments, group them into a hash and pass that instead.

This means we don't need to name each argument, as the hash key does that instead. It also reduces the number of arguments being passed in.

* Use do/end for multiline blocks, not curly braces