Update CanStripWhitespace to work with frozen strings (#345)
* Update CanStripWhitespace to work with frozen strings

In Ruby 3, all string literals are immutable. This isn't true in Ruby 2, but to prepare for an eventual upgrade, we are adding the comment `# frozen_string_literal: true` to the start of all our files. Rubocop enforces this.

However, while adding this comment, we discovered the `string.strip!` command in our CanStripWhitespace concern violates this, as it attempts to modify a frozen string.

This PR refactors the concern to work with frozen strings.

* Reassign key value instead of using bang method

Rather than just call `strip!` on the now-immutable value, we return the result from `strip_string` and reassign the value of the key.