Format postcode without modifying frozen strings (#347)
In Ruby 3, all string literals are immutable. This isn't true in Ruby 2, but to prepare for an eventual upgrade, we are adding the comment # frozen_string_literal: true to the start of all our files. Rubocop enforces this.

However, while adding this comment, we discovered that the `strip!` and `upcase!` commands in PostcodeForm violate this, as they attempt to modify a frozen string.

This PR refactors our postcode forms to work with frozen strings.