Fix style issues in form specs (#346)
* Fix style issues in form specs

This PR fixes issues flagged by Rubocop in our form specs. Issues in forms themselves will be dealt with in a separate PR to keep file numbers manageable.

Part of our ongoing housekeeping.

* Add frozen_string_literal comment

* Remove unnecessary assignments when using shared examples

We use shared examples to validate some form fields. These examples take the name of the form and sometimes the field name. Previously we assigned these as we passed them in as arguments, but this doesn't do anything so should be removed.

* Use do-end for multiline blocks

Smart answers forms were using multiline blocks with curly braces. Bad.