Refactor tier decision for smart answers (#353)
* Refactor tier decision for smart answers

This change is part of our ongoing housekeeping to enable Rubocop as part of our CI.

Rubocop flagged that the complexity of the method we used to check if a registration should be upper or lower tier based on the smart answers was high. So this PR attempts to solve this issue and make the code more readable.

* Create SmartAnswersCheckerService

This service exists to check the responses to smart answers and return decisions based on that - in this case, whether a registration should become lower tier.

* Use service instead of keeping logic in concern

This removes some complexity from CanChangeWorkflowStatus.

* Move logic into separate methods to reduce complexity

Previous changes had only moved the complex method, rather than fixed it.

This commit splits the logic into multiple methods so we no longer have 8 checks happening in one method.