Fix translation error on main people form (#432)
Spotted a translation error following our refactoring in https://github.com/DEFRA/waste-carriers-engine/pull/428, as part of the main people form had been moved to a partial. Updating the language YAML file fixed this.

However, this fix had a knock-on effect of breaking the "Add another person" link, which relied on a specific param being submitted based on the YAML file.

Fixing this then broke the "Add another person" on the conviction details form, which did not have a partial and so had a differently-structured YAML file. Sooo I moved this to a form partial as well and now everything works again. (Phew.)

Obviously this is a clear indicator we need to do some refactoring about how this controller works! So here's a quick fix for now but I'll log an issue to come back to this.

Our request tests should be raising translation errors so they don't make it to master. This PR also updates the dummy app config to raise the errors in future.