Fix Rake (#443)
PR #355 and PR #354 fixed a series of issues having applied our latest and greatest [defra-ruby-style](https://github.com/DEFRA/defra-ruby-style) rules.

Part of that was changing some calls to `__FILE__` to use `__dir__` instead. Unbeknownest to us this actually broke rake and we only came to realise this when attempting to update the CHANGELOG.

In the dummy app's `Rakefile` the line was

```ruby
require File.expand_path("../config/application", __FILE__)

require File.expand_path("../config/application", __dir__)

require File.expand_path("config/application", __dir__)
```

Now this change could have just corrected the line and moved on. However a check of our other engines in WEX and FRAE showed that they were setup differently. That is to be expected because WEX and FRAE were always engines, however WCR started out as an app we converted to an engine.

So this change brings both the engine and the engine's dummy app Rakefiles inline with our other engines (TLDR; leave the dummy app as it gets generated, completely re-write the engine Rakefile 😁).