Use shared phone number validator (#457)
This PR removes our built-in phone number validation in favour of the shared validator in defra-ruby-validators.

We also use the gem validator to clean data. There is a method used to remove old phone numbers which may have been entered using the older version of the app and did not meet our current data standards.