Use shared Companies House number validator (#459)
This PR removes our built-in company_no validation in favour of the shared validator in defra-ruby-validators.

One difference in behaviour here is that the gem validator does not automatically pass if the transient_registration being validated doesn't need a `company_no`. So instead we just only call the validator when we know the value is required.

In the case of the registration number form, a user should not be on it unless the number is required, so we just skip the check.

* Add config for Companies House validation

This also introduces `WasteCarriersEngine.configuration`, which is a pattern we started following in Waste Exemptions and intend to use more here. As a first pass, we just use it to provide the necessary config for the Companies House validation in DefraRuby::Validators.