Forms refactoring: Submit with assign and validation (#486)
* Forms refactoring: Submit with assign and validation

From: https://eaflood.atlassian.net/browse/RUBY-693

As a start for the form refactoring, to get the code in-line with what we have in WEX, the goal of this PR is to get the `submit` method in the `BaseForm` to use a mass-assignment strategy on the transient registration so that we can be able to both `delegate` or `assign` attributes.
The second parameter passed to the `submit` value has been killed, a refactoring will follow remiving the parameter to be passed by the children classes later. I see no point on that parameter to exists, and when / if we find that we need to assign a reg_identifier we can take advantage of mass-assignment like we will do with the rest of the parameters.
Changes on the tests are missing consistency. There are different implementations of the form specs, some uses a unit tests strategy, some other uses an integration tests strategy. I have stubbed test where the validation failure was failing due to the form having no validations on his own. There is a case where the validation were tested in isolation in a more unit-style. After discussing it with Alan we agreed to fix stuff as we can while we work on the forms refactoring, then group together and decide how we want our test suite to be structure and subsequentially come back to this specs and bring some consistency to them. This work is already captured as its own story for next sprint.

* Fix rubocop

* Fix failing spec