Forms Refactoring - Address lookup forms (#514)
* Forms Refactoring - Address lookup forms

From: https://eaflood.atlassian.net/browse/RUBY-693

Apply refactoring to Lookup forms.
The forms are now sending a `uprn` through the select field, which is then used to match the correct address with the list of temp_addresses from OS places. This ensure we can use standard Rails strategy of Strong Parameters and mass assign.
Since the assignment of contact and company addresses is done through a non-standard rails association, what is different from WEX is that in this case we will pass to the `transient_registration` an instance of `Address` rather than `address_attributes`. This is due to the fact that we cannot use `accept_nested_params` for a fake association.