Improve metadata status checks (#520)
This PR cleans up our status checks on registrations and transient_registrations so we can call something like `registration.active?` instead of `registration.metaData.ACTIVE?` or `registration.metaData.status == "ACTIVE"`.

Sometimes we just want to get and output the status. Now we can call `registration.status` and get a nice, neat output.

This was suggested in reviews of https://github.com/DEFRA/waste-carriers-back-office/pull/387

This PR also adds the `INACTIVE` status. It shouldn't be possible to set this status through the engine yet, but we do have existing registrations that have it set, so we should account for its existence.