Add pending_manual_conviction_check? method to registration (#535)
https://eaflood.atlassian.net/browse/RUBY-704

This may seem redundant, but it means that registrations and transient_registrations have the same interface, which will simplify the checks for the back office dashboard.

This PR also makes our reg attribute shared examples work with multiple models. These were always using a transient_registration to test with, when we should be specifying the factory as an argument.