Update EntityMatchingService to call new convictions check (#556)
https://eaflood.atlassian.net/browse/RUBY-741

Java be gone! This PR swaps out our old calls to the Java service in favour of calling our new conviction match services instead.

Because the data we are returning is slightly nicer (no weird Unix time in seconds), there may be some knock-on effects in other classes too. This is also an opportunity to further implement BaseService.

This should make several cassettes redundant.