Add shared convictions scopes for regs and transient_regs (#572)
https://eaflood.atlassian.net/browse/RUBY-791

These are for use on the convictions dashboard.

There was a previous version which only applied to transient registrations, but that also included a `submitted` scope. I think we should just apply this separately where we use it (which is only on the current version of the convictions dashboard, as far as I can tell).