Add service that generates Base(58) secure tokens (#580)
In WEX and FRAE we have relied on using [HasSecureToken](https://github.com/robertomiranda/has_secure_token) to generate unique random tokens of a fixed length in the objects we create.

We have a need to uniquely identify a record but cannot use its ID because it also has to be something 'unguessable'.

HasSecureToken works for WEX and FRAE because it is designed to work with ActiveRecord models. WCR on the overhand uses MongoDb, which means we can't use it.

So instead, we've copied it wholesale into our project as a service we can call when we need to generate a token. Thanks, @robertomiranda 😁!