Set up RegistrationCompletionService (#583)
https://eaflood.atlassian.net/browse/RUBY-793

When conviction checks for new registrations are moved into the back office, this is the first time we'll need to 'complete' a new registration in the new codebase. So this PR creates a RegistrationCompletionService to handle that.

When the new registration journey starts to use transient_registrations, this will probably need some modifying to deal with those instead of the registration model.