Persist setting of expiries_on when activating (#584)
https://eaflood.atlassian.net/browse/RUBY-793

Spotted that while the `activate` event sets the `expires_on` in memory, it doesn't persist it. This means that if the registration was assigned an `expires_on` when it was first submitted through the frontend, that value sticks, although the behaviour we want is for it to expire 3 years after approval.

So this fixes that.