Use secure token in path instead of reg. identifier (#579)
We spotted when working on implementing the ability to order copy cards in the back-office that we would face an issue trying to reuse forms where the expectation was to find a transient registration using just the registration identifier.

For example, CBDU100 is an active registration. I then start a renewal so create a `RenewingRegistration` which sits in the `transient_registrations` collection. But then I need to take a copy card order for the existing active registration. That would create an `OrderCopyCardsRegistration` transient registration with the same registration identifier of CBDU100.

So how would the renewing journey or the order copy cards journey know which transient registration to pull if we only have the reg_identifier as an ID?

We solved this in WEX by having all transient registrations automatically creating a `token` (which is a Base58 24 character string) when first saved to the DB, and then using that in the URL to identify the record.

So this change, which touches loads of files in the codebase, actually is about switching to using the token rather than the registration identifier in our journeys.

It also brings the URL pattern inline with WEX, so `/{token}/stage/`. This resolves an issue we had in WEX and WCR with validations losing the ID (because it was at the end of the URL rather than the beginning) which then broke going back after a validation error.