Send confirmation email when registration is completed (#585)
https://eaflood.atlassian.net/browse/RUBY-793

When a registration is completed, we should send a confirmation email to the user.

This should be the same as the existing email sent by the frontend. (The email itself has been tidied up for ease of use, but the result should be the same.)

For now, the only time this will happen is when it's approved after a convictions check.