Set up EditRegistration model (#661)
https://eaflood.atlassian.net/browse/RUBY-827

This PR sets up a new EditRegistration model to be used for editing. This will be a new type of TransientRegistration.

An EditRegistration should copy data from an existing Registration when it is created. It will also have its own workflow state machine.

This PR also adds a CanCopyDataFromRegistration concern, which is used for both EditRegistrations and RenewingRegistrations. It does need to behave slightly differently for both so should be configured in each model.