Set up edit overview page (#666)
https://eaflood.atlassian.net/browse/RUBY-829

We need a new 'edit overview', as per wireframe, page that acts as the core of the edit journey. This page shows information about the registration and has links to change information.

I also delete all the SCSS files that were sitting around. None of these are imported to the host apps, which have their own copies of them. So might as well just delete them to avoid any confusion.