Fix view certificate for lower tier registrations (#670)
https://eaflood.atlassian.net/browse/RUBY-834

We have recently worked on adding the ability to see the certificate we generate when a registration is complete from the back office. However, in QA it was found that an error occurs when you try and use the link for a lower-tier registration.

Upon investigation, we found this error in the logs

```bash
I, [2020-02-12T14:26:56.768579 #27414]  INFO -- :   Rendered /srv/ruby/waste-carriers-back-office/shared/bundle/ruby/2.4.0/bundler/gems/waste-carriers-engine-e39f56f6bac0/app/views/waste_carriers_engine/pdfs/certificate.html.erb (3.8ms)
I, [2020-02-12T14:26:56.769213 #27414]  INFO -- : Completed 500 Internal Server Error in 11ms
F, [2020-02-12T14:26:56.776346 #27414] FATAL -- :
ActionView::Template::Error (undefined method `in_time_zone' for nil:NilClass):
    125:                 </tr>
    126:                 <tr>
    127:                   <td class="table-column-text"><%=t ".expiry_date" %></td>
    128:                   <td><%= @presenter.expires_on.in_time_zone("London").to_date %></td>
    129:                 </tr>
    130:               </tbody>
    131:             </table>
  app/controllers/certificates_controller.rb:10:in `show'
```

Essentially the view always assumed that it was dealing with an upper-tier registration, which to be fair till this point it had. As lower tier registrations don't have an expiry date, an error would be thrown.

In the same vein, we also spotted that the **Registered as** section had a missing translation. So we needed to update the presenter and locals to also work with lower-tier registrations