Don't send confirmation email for fully AD copy card orders (#675)
https://eaflood.atlassian.net/browse/RUBY-896

When additional copy cards are ordered, if there is no contact email or it's the default NCCC inbox, we should not send any sort of confirmation email.

To do this, we add assisted_digital_email to the engine config. This env var is what we already use to set the account_email in the frontend, so it should already be set on all environments.