Fix methods checking whether a reg is overseas (#687)
I found this issue when I created a new overseas registration in the frontend for testing purposes.

I noticed that while I could set the location to "overseas", I still saw the business type screen. In the engine, overseas regs skip this form and set the business type automatically, but this isn't the case in the frontend. There isn't an `overseas` option on this form so you have to select something else.

So the registration I created had a `location` of `overseas`, but a `business_type` of `limitedCompany`.

However, this caused problems when I loaded it into the engine to test the edit feature. When I tried to edit the address, it showed me the UK version of the form, with the wrong postcode behaviour and no country field:

This turned out to be because we had two methods in CanHaveRegistrationAttributes called `overseas?` (which checked the `location` attribute) and `business_is_overseas?` (which checked the `business_type` attribute).

This meant that I could create a registration which returned `true` for `overseas?` but `false` for `business_is_overseas?`.

This isn't a problem when the manual address form is loaded as part of the renewal process, as users have to go through the location form first and it updates the `business_type` to `overseas` as well, making the response from the two methods consistent.

But when I loaded the form as part of an edit, it called `business_is_overseas?` on my registration, got a `false` and rendered the UK version of the form.

To fix this, I've merged the functions so we just have `overseas?` returning true if either the location or the business_type are the correct match. This should cover all the possible versions of an overseas registration:

- registrations created before the location form was added, which would have a `location` of `nil` and a `business_type` of `overseas`
- registrations created after the location form was added, which would have a `location` of `overseas` and a `business_type` of some other value
- registrations which have been renewed, and so have a `location` and `business_type` both set to `overseas`
- and, for the future, registrations created in the front-office, which also have a `location` and `business_type` both set to `overseas`

I think it might be worth trying to clean up our data, but probably not until the frontend is decommissioned and we know any registrations we create will be consistent.