Complete an edit (#686)
https://eaflood.atlassian.net/browse/RUBY-831

Completing an edit should update the registration with the new data and delete the old EditRegistration record.

This PR covers setting up the 'edit complete' page and making sure the edit actually completes as specified.

Also tweaked the registration factory has finance_details by default. This is more accurate to what a real upper tier reg would look like and was necessary for properly testing the edit completion. This slightly broke some of the copy card tests, so we had to fix those as part of the PRs.