Store a past_registration when completing an edit (#698)
https://eaflood.atlassian.net/browse/RUBY-831

When a user makes an edit to a registration, we should store a new past_registration object containing the pre-edit version of the registration.

This will let us keep track of changes made by the edit.

This should happen at the point of completion. Cancelling an edit does not save a past_registration.

If a past registration is created due to an edit, we store this in a new `cause` field.

We also use this to make sure we always save past regs when they're because of an edit, as previously this could be blocked if there was already an existing one with the same expiry date.