Don't check convictions when doing an edit (#699)
Submitting the declaration form runs the conviction checks by default. This is fine for renewals and new registrations. However the edit workflow also uses this form, but we don't want to run the conviction checks during editing.

This PR makes a small tweak to only check convictions for particular journeys.