Add start form for new registrations (#695)
https://eaflood.atlassian.net/browse/RUBY-840

This adds a new page to the service which allows a user to start the new registration journey.

The page is part of the new registration workflow.

The next pages after this are 'location' and 'renew a registration'. Location as a placeholder already existed. So we also add a placeholder for 'renew a registration' as part of this change. This allows us to test and confirm the behaviour of the start form.