Hide start new registration behind feature toggle (#718)
https://eaflood.atlassian.net/browse/RUBY-840

We are currently not yet finished with building new registrations into the service (migrating it from the old code). But there are other features we do want to ship so for now, we will hide the ability to start a new registration behind a feature toggle.

** Constrain the start of the journey

We figured the easiest and most effective way to stop access to the journey was to constraint access to the start form. If you can't access the start and create your new registration, you should be blocked from the whole thing.