Calculate Reg identifier for new registrations on the fly (#732)
* Calculate Reg identifier for new registrations on the fly

We have decided that we will assign a unique registration number to a registration when the declaration form is submitted. However, since the user is able to go back in the journey and change its tier up until the very end of the journey, we also decided that would be wise to not persist the `CBDL` VS `CBDU` prefix to the identifier until the transient object becomes an actual registration. With this change, we always generate the number on the fly for NewRegistration objects.