Build finance_details object for edit journey (#737)
https://eaflood.atlassian.net/browse/RUBY-833

This is part of a set of multiple PRs to configure payment when a carrier type changes during editing.

This PR sets up a BuildEditFinanceDetailsService class which is used to create finance_details objects for the edit journey. It's pretty similar to OrderAdditionalCardsService so there's probably room for consolidating once we have a consistent way of building this objects (I'm looking at you, renewals).

This PR also renames the former OrderAdditionalCardsService to BuildOrderCopyCardsFinanceDetailsService. This makes the two services generating finance_details have more consistent names. It also makes the purpose of this service clearer as there are multiple services dealing with ordering additional copy cards.