Pay for a new registration (#738)
From: https://eaflood.atlassian.net/browse/RUBY-837

This PR implements the code that prepares a new registration for payment. It will use a service to build the correct finance details and assign it to the registration.
It also contains some refactoring arising from the need to deal with user email's rather than user objects, as in the case of new registration we won't have a current_user to work with.