Set up workflow for edit payments by bank transfer (#741)
https://eaflood.atlassian.net/browse/RUBY-833

This is part of a set of multiple PRs to configure payment when a carrier type changes during editing.

This PR sets up the workflow and pages for paying for an edit by bank transfer. This includes the payment summary and bank transfer info pages. It also includes several shared partials which can hopefully be reused on similar pages.

WorldPay will be added separately in a future PR.