Backfill individual DOB fields when initialising key person (#756)
https://eaflood.atlassian.net/browse/RUBY-830

We spotted that when editing a sole trader's key person, the birthday wasn't always pre-populated.

This turned out to be because any registrations generated in the frontend and not since modified did not store values for `dob_day`, `dob_month` and `dob_year`. Instead they just had the single `dob`. But any records generated in the new engine have all 4 values.

To get around this, we set the values of the individual fields to match the value of `dob`. This happens when a new key person is initialised, specifically when we load an older one from the database.