Ask for business type when registration is edited from overseas to UK (#759)
https://eaflood.atlassian.net/browse/RUBY-830

We spotted an issue where if a user edited a registration's location, but it had previously been overseas, the business type would still be "overseas". This was causing some problems.

Instead we want to get a more accurate business type for this now-UK business. So after changing the location from overseas to a UK location, we will immediately ask for the business type instead of redirecting to the edit overview.