Location cannot be edited (#764)
https://eaflood.atlassian.net/browse/RUBY-830

The ability to edit location was causing lots of problems for us, as switching between overseas and non-overseas affects the other data that you are asked to enter. This meant we could potentially get incomplete data.

Unfortunately all our workarounds (like asking for business types as well) just opened up more potential gaps. So we've decided it's better to just not have the ability to edit locations for now.

Since the purpose of the location question is mainly to filter out people during the registration/renewal process, it has limited relevance during editing anyway.