Make main people form work with old business types (#765)
We still have some regs hanging around with `publicBody` or `authority` business types.

The main people form was designed for the renewals journey, where these no longer exist. So it needs a few tweaks to work with these old types, just around text and the validation for age limits.

This PR simplifies the age validation while we're at it.