Add renewal number lookup page (#767)
https://eaflood.atlassian.net/browse/RUBY-841

When external users enter the service, they see a page that asks what the user wants to do - start a new registration or renew. If they select 'renew', they see this page: 'What is your waste carrier registration number?'

We need to replace this page from the frontend with a new version.

This PR adds a temp_lookup_number attribute. This will be used to lookup existing registration numbers, but it's technically part of the new reg journey.

In the controller, we have to diverge from the usual submit method, as we don't want to call `next!` on the transient_registration, since we are actually handing over to a new transient_registration instead (`new` to `renewing`).