Send email when user registers and has pending payment (#786)
https://eaflood.atlassian.net/browse/RUBY-910

If a user submits a new registration and chooses to pay by bank transfer, we should email them with the bank details and confirm that their registration won't be activated until payment is received.

One highlight is we have formatted the email link such the Reg number automatically gets included in the subject and body. Neat!