Factory-generated renewals always initialise with registration (#811)
Without this, the renewal object cannot validate, which essentially means that we have to include the `:has_required_data` trait any time we want to create a renewal. This is a huge pain - it clutters up the specs and also makes it difficult to reuse some shared examples.

I plan to do an additional sweep of the specs once this is merged and remove no-longer-needed `:has_required_data` calls.