Add 'are you sure?' screen when paying by bank transfer (#820)
https://eaflood.atlassian.net/browse/RUBY-1004

Users are confused by the current bank transfer flow. An 'are you sure?' page has been added to help users pay by bank transfer.

This replaces the existing bank transfer information page. All details for how to pay by bank transfer will be moved to the page at the end of the journey as part of RUBY-1005.

This only applies to the new registration and renewals journey. Edit and ordering copy cards are not affected.

As part of this, we rename BankTransferForm to ConfirmBankTransferForm. We also stop the CopyCardsBankTransferForm from inheriting, since it will no longer have the same delegated attribute.