Fix titles rendering on error renew pages (#852)
These pages are rendered using the `new` action but different templates name. For this reason, the usual method of rendering titles which is present in the application helper cannot fetch the correct translation, as it is only based on the action name.
Hence, we decided to explicitely set titles using `content_for`.