Revert addition of payment receipt screen (#889)
https://eaflood.atlassian.net/browse/RUBY-1141

In [PR #876](https://github.com/DEFRA/waste-carriers-engine/pull/876) we added a new screen to the journey, between copy cards and payment summary. This would allow a user to enter the email address they wished their payment summary to be sent.

This was our initial stab at solving the issue of removing the account email from the new 'new registration' journey. The existing version of the journey forwards te account email to Worldpay when paying by card. Because the email is passed on Worldpay will then send a payment confirmation to that email address.

By removing the account email, it was our intent to just forward the contact email to Worldpay. However, it was highlighted this was likely to be a cause for complaint from those who act as agents because it would expose the actual cost of a registration. This is because in our contact email screen we specifically ask for the email of the registrant, not the person filling in the form.

The issues with this solution were

- it referred to a receipt when actually it's just a payment confirmation
- it asks all users to complete the field, even though only those paying by Worldpay will receive the confirmation

So our UCD team have come back with an improved version which we are ready to implement. First step though is to revert the change made in PR #876 and put the journey back how it was.

We'll keep the new receipt_email field as we will still be using that as well the logic to pick which email address to pass to Worldpay. But we will delete the receipt email form page and associated code and tests plus update the journeys to remove it as a step.