Hide receipt email on payment summary unless WorldPay selected (#892)
https://eaflood.atlassian.net/browse/RUBY-1141

In a truly shocking development, this adds some JavaScript to the service.

The purpose of this is to hide the receipt email field from the payment summary page until the "WorldPay" option is selected.

If the page is loaded and "WorldPay" is already selected (eg if the page has been reloaded due to a validation error) then the email field will still be shown.

If "bank transfer" is later selected instead, the field will be hidden again.

Note that this won't work in IE8 and below. The default for these browsers and for disabled JavaScript is to always display the email field.

We will also need to add a little `<%= yield :page_scripts %>` to the application template on the host apps, so I'll sort that out separately.