Hide payment confirmation in 'end' pages (#897)
https://eaflood.atlassian.net/browse/RUBY-1141

We have come across an issue where we need to change behaviour in the journey based on whether a registration or renewal is being carried out in either the front-office app or the back-office app.

We should have known, but have just realised that Worldpay does not send payment confirmation emails if the merchant code used is MOTO.

So we shouldn't show content in the back-office that suggests a payment confirmation email will be sent. This change hides this content in the 'end' pages.

N.B. Also includes an update to content in Worldpay pending payment page. Whilst working on this change we noticed that the content between this page and the pending conviction page differed.