Add payment confirmation email to Edit (#891)
https://eaflood.atlassian.net/browse/RUBY-1165

To keep the edit view consistent with the registration details view this change adds the Card payment confirmation to the edit view.

Unlike the registration details view though, the section will always display even if the `:receipt_email` field is not present or empty. This is to remain consistent with how we treat the account email in the edit view.