Generate renewal token when one is requested (#898)
https://eaflood.atlassian.net/browse/RUBY-1166

If a registration is eligible for renewal NCCC would like us to display the same renewal link that gets included in the renewal reminder emails and letters in registration details.

Currently, though we only generate the token used in the link when first generating the reminder email. A             registration is eligible for renewal before this is done. And when we first release there will be hundreds that are eligible which will never go through the reminder process because it is too late.

The simplest solution we have for this is to generate it if required. This way we avoid any timing issues or another scheduled task that would have to try and identify all eligible registrations and update them in one go.

So this change will add some logic to automatically generate and save a token if one does not exist when the token is requested.

The key thing to note is to access the underlying data field you need to use `self[:field_name]`. So now we can just ask for `renew_token` where we need it, and one will be generated automatically if it doesn't exist.

This also means `generate_renew_token!()` is no longer needed so we remove it in this change.

N.B Remove dummy app .ruby-version file

Kept getting messages about using the wrong version of bundler when trying to run up the rails console in the dummy app.

Spotted that we had a .ruby-version file with the old version in it. I could have corrected the version but actually there is no need for it to be here when we have one in the root of the project.