Remove magic link constraint for renew endpoint (#900)
https://eaflood.atlassian.net/browse/RUBY-1170

RUBY-1170 merged `email_reminders` and `renew_via_magic_link` into a new single feature toggle called `renewal_reminders`. Our work has come so far that we don't actually make a distinction between renewal reminders and renewal reminders with magic links. There is only the version with magic links.

So we needed to update the engine to use the new toggle, but instead have removed it.

If we continued to use the toggle against the `/renew` endpoint the following scenario would happen

- renewal reminders are enabled
- user (along with others) gets a renewal reminder email containing a feature toggle
- NCCC request renewal reminders are disabled because of an increase in calls
- user then decides to try and renew, clicks link in the email but gets an error
- user calls NCCC to report issue along with everyone else trying!

So we will offer a better experience if the `/renew` endpoint remains open. It is only of use if you have a valid token, and we'll only be generating those if `renewal_reminders` is enabled.