Hide contact name, phone and email in renewals (#901)
https://eaflood.atlassian.net/browse/RUBY-1171

NCCC want to allow as many users as possible to renew via magic link. 2 major blockers are

- 3rd party agents acting on behalf of users
- the 3 year gap between renewing where contacts at an organisation have changed

This means the automated renewal reminders containing magic links won't get to who they need to. When these people call NCCC to renew (because they didn't receive the reminder) NCCC would like to manually send them a registrations renewal link. We've already covered exposing the link in [PR #963](https://github.com/DEFRA/waste-carriers-back-office/pull/963). But to ensure we don't inadvertently expose any personal details NCCC would also like the contact details which are not currently public, to not pre-populate for renewals.

Ideally we would have 2 journeys; one for recognised users with as much information as possible pre-populated, and one for anonymous renewals. But with our current resources reducing the amount we pre-populate will help more people renew online more quickly in the short-term.

So this change stops the following being copied into the renewal when initialised from a registration

- contact name
- contact phone
- contact email

* Implement change and add new test

This makes the change we need. But it does cause the tests around phone number to fail. Making the change separately though to make review a little clearer.

* Remove phone number tests

We now always don't copy the phone number so don't need the tests for handling valid and invalid phone numbers.

* Delete remove_invalid_attributes() from concern

`remove_invalid_attributes()` covered both the phone number and revoked reason. With handling the phone number no longer necessary that just leaves the revoked reason. And the only reference to `remove_invalid_attributes` is in the `RenewingRegistration` initializer.

So this change is some refactoring to remove now redundant code.

* Update renewing registration factory

Now we don't copy across contact details from the registration we need to specify them in the factory when the traits `:has_required_data` and `:has_required_overseas_data` are used. Else a lot of tests start failing!

* Update worldpay fixture