Don't copy account email around editing (#902)
https://eaflood.atlassian.net/browse/RUBY-1178

https://eaflood.atlassian.net/browse/RUBY-1151

Previously during an edit, the value of account_email was copied from the registration to the transient_registration and then copied back again when the edit was completed.

If the registration was transferred between the edit being started and completed, this resulted in a bug where the account_email date was "reset" to its previous value.

There was existing code to try and prevent this. But there was a mistake in the naming convention used caused by a legacy issue of the old app not being consistent with camelCase and snake_case field names.

This PR fixes the existing code.