Send 'renewal received pending conviction check' emails from Notify (#988)
https://eaflood.atlassian.net/browse/RUBY-1392

Emails are sent on a change of state.

The logic to determine which template to use was in the ReminderMailer. This
won't work any more, as we are replacing the mailer with a single-responsibility
Notify class.

Here we specify the notify service/template in the state machine.
I hope it works.