Adding 'Error' in the page title when an error is generated (#1088)
https://eaflood.atlassian.net/jira/software/projects/RUBY/boards/374?selectedIssue=RUBY-1726

Here we have updated the page titles to change when an error is generated. This allows screen readers to give a better description to the user on the pages status.

The titles will now read <% Error - Registration start - Register as a waste carrier - GOV.UK %> when a error has been generated.
One test has been written in application_helper_spec to check the changes.